# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace API",
    "description": "Marketplace API",
    "version": "1.0",
    "contact": {
      "name": "Marketplace",
      "url": "https://ordercloud.io",
      "email": "oheywood@four51.com"
    }
  },
  "servers": [
    {
      "url": "https://marketplace-api-qa.azurewebsites.net",
      "description": "Marketplace API"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/oauth/token",
            "scopes": {
              "AddressAdmin": "",
              "AddressReader": "",
              "AdminAddressAdmin": "",
              "AdminAddressReader": "",
              "AdminUserAdmin": "",
              "AdminUserGroupAdmin": "",
              "AdminUserGroupReader": "",
              "AdminUserReader": "",
              "ApiClientAdmin": "",
              "ApiClientReader": "",
              "ApprovalRuleAdmin": "",
              "ApprovalRuleReader": "",
              "BuyerAdmin": "",
              "BuyerImpersonation": "",
              "BuyerReader": "",
              "BuyerUserAdmin": "",
              "BuyerUserReader": "",
              "CatalogAdmin": "",
              "CatalogReader": "",
              "CategoryAdmin": "",
              "CategoryReader": "",
              "CostCenterAdmin": "",
              "CostCenterReader": "",
              "CreditCardAdmin": "",
              "CreditCardReader": "",
              "FullAccess": "",
              "IncrementorAdmin": "",
              "IncrementorReader": "",
              "InventoryAdmin": "",
              "MeAddressAdmin": "",
              "MeAdmin": "",
              "MeCreditCardAdmin": "",
              "MessageConfigAssignmentAdmin": "",
              "MessageSenderAdmin": "",
...
    "schemas": {
...
      "CatalogXp": {
        "type": "object",
        "example": {},
        "properties": {}
      },
...
      "Inventory": {
        "type": "object",
        "example": {"Enabled":false,"NotificationPoint":0,"VariantLevelTracking":false,"OrderCanExceed":false,"QuantityAvailable":0,"LastUpdated":"2018-01-01T00:00:00-06:00"},
        "properties": {
          "Enabled": {
            "type": "boolean"
          },
          "NotificationPoint": {
            "type": "integer",
            "format": "int32"
          },
          "VariantLevelTracking": {
            "type": "boolean"
          },
          "OrderCanExceed": {
            "type": "boolean"
          },
          "QuantityAvailable": {
            "type": "integer",
            "format": "int32"
          },
          "LastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MarketplaceCatalog": {
        "type": "object",
        "example": {"xp":{},"ID":"","OwnerID":"","Name":"","Description":"","Active":false,"CategoryCount":0},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CatalogXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "OwnerID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          },
          "Active": {
            "type": "boolean"
          },
          "CategoryCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MarketplaceCatalogAssignment": {
        "type": "object",
        "example": {"ID":"","CatalogID":"","BuyerID":"","ViewAllCategories":false,"ViewAllProducts":false},
        "properties": {
          "ID": {
            "type": "string"
          },
          "CatalogID": {
            "type": "string"
          },
          "BuyerID": {
            "type": "string"
          },
          "ViewAllCategories": {
            "type": "boolean"
          },
          "ViewAllProducts": {
            "type": "boolean"
          }
        }
      },
...
      "MarketplacePriceSchedule": {
        "type": "object",
        "example": {"xp":{},"ID":"","Name":"","ApplyTax":false,"ApplyShipping":false,"MinQuantity":0,"MaxQuantity":0,"UseCumulativeQuantity":false,"RestrictedQuantity":false,"PriceBreaks":[{"Quantity":0,"Price":0}]},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriceScheduleXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "ApplyTax": {
            "type": "boolean"
          },
          "ApplyShipping": {
            "type": "boolean"
          },
          "MinQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "MaxQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "UseCumulativeQuantity": {
            "type": "boolean"
          },
          "RestrictedQuantity": {
            "type": "boolean"
          },
          "PriceBreaks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceBreak"
            }
          }
        }
      },
      "MarketplaceProduct": {
        "type": "object",
        "example": {"xp":{"Images":[{"URL":"","Tag":{"VariantID":"","Options":[""]}}],"Status":"Draft","HasVariants":false,"Note":"","Tax":{"Category":"","Code":"","Description":""},"UnitOfMeasure":{"Qty":0,"Unit":""},"ProductType":"Standard"},"OwnerID":"","DefaultPriceScheduleID":"","AutoForward":false,"ID":"","Name":"","Description":"","QuantityMultiplier":0,"ShipWeight":0,"ShipHeight":0,"ShipWidth":0,"ShipLength":0,"Active":false,"SpecCount":0,"VariantCount":0,"ShipFromAddressID":"","Inventory":{"Enabled":false,"NotificationPoint":0,"VariantLevelTracking":false,"OrderCanExceed":false,"QuantityAvailable":0,"LastUpdated":"2018-01-01T00:00:00-06:00"},"DefaultSupplierID":""},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductXp"
              }
            ]
          },
          "OwnerID": {
            "type": "string"
          },
          "DefaultPriceScheduleID": {
            "type": "string"
          },
          "AutoForward": {
            "type": "boolean"
          },
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          },
          "QuantityMultiplier": {
            "type": "integer",
            "format": "int32"
          },
          "ShipWeight": {
            "type": "number",
            "format": "float"
          },
          "ShipHeight": {
            "type": "number",
            "format": "float"
          },
          "ShipWidth": {
            "type": "number",
            "format": "float"
          },
          "ShipLength": {
            "type": "number",
            "format": "float"
          },
          "Active": {
            "type": "boolean"
          },
          "SpecCount": {
            "type": "integer",
            "format": "int32"
          },
          "VariantCount": {
            "type": "integer",
            "format": "int32"
          },
          "ShipFromAddressID": {
            "type": "string"
          },
          "Inventory": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Inventory"
              }
            ]
          },
          "DefaultSupplierID": {
            "type": "string"
          }
        }
      },
      "MarketplaceProductAssignment": {
        "type": "object",
        "example": {"ID":"","ProductID":"","BuyerID":"","UserGroupID":"","PriceScheduleID":""},
        "properties": {
          "ID": {
            "type": "string"
          },
          "ProductID": {
            "type": "string"
          },
          "BuyerID": {
            "type": "string"
          },
          "UserGroupID": {
            "type": "string"
          },
          "PriceScheduleID": {
            "type": "string"
          }
        }
      },
      "MarketplaceProductFacet": {
        "type": "object",
        "example": {"xp":{"Options":[""],"ParentID":""},"ID":"","Name":"","XpPath":"","ListOrder":0,"MinCount":0},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductFacetXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "XpPath": {
            "type": "string"
          },
          "ListOrder": {
            "type": "integer",
            "format": "int32"
          },
          "MinCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MarketplaceSpec": {
        "type": "object",
        "example": {"xp":{"UI":{"ControlType":"Text"}},"ID":"","ListOrder":0,"Name":"","DefaultValue":"","Required":false,"AllowOpenText":false,"DefaultOptionID":"","DefinesVariant":false,"OptionCount":0,"Options":[{"ID":"","Value":"","ListOrder":0,"IsOpenText":false,"PriceMarkupType":"NoMarkup","PriceMarkup":0}]},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpecXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "ListOrder": {
            "type": "integer",
            "format": "int32"
          },
          "Name": {
            "type": "string"
          },
          "DefaultValue": {
            "type": "string"
          },
          "Required": {
            "type": "boolean"
          },
          "AllowOpenText": {
            "type": "boolean"
          },
          "DefaultOptionID": {
            "type": "string"
          },
          "DefinesVariant": {
            "type": "boolean"
          },
          "OptionCount": {
            "type": "integer",
            "format": "int32"
          },
          "Options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecOption"
            }
          }
        }
      },
      "MarketplaceSpecOption": {
        "type": "object",
        "example": {"xp":{"Description":"","SpecID":""},"ID":"","Value":"","ListOrder":0,"IsOpenText":false,"PriceMarkupType":"NoMarkup","PriceMarkup":0},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpecOptionXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Value": {
            "type": "string"
          },
          "ListOrder": {
            "type": "integer",
            "format": "int32"
          },
          "IsOpenText": {
            "type": "boolean"
          },
          "PriceMarkupType": {
            "enum": [
              "NoMarkup",
              "AmountPerQuantity",
              "AmountTotal",
              "Percentage"
            ],
            "type": "string"
          },
          "PriceMarkup": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "MarketplaceSpecProductAssignment": {
        "type": "object",
        "example": {"ID":"","SpecID":"","ProductID":"","DefaultValue":"","DefaultOptionID":""},
        "properties": {
          "ID": {
            "type": "string"
          },
          "SpecID": {
            "type": "string"
          },
          "ProductID": {
            "type": "string"
          },
          "DefaultValue": {
            "type": "string"
          },
          "DefaultOptionID": {
            "type": "string"
          }
        }
      },
...
      "PriceBreak": {
        "type": "object",
        "example": {"Quantity":0,"Price":0},
        "properties": {
          "Quantity": {
            "type": "integer",
            "format": "int32"
          },
          "Price": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "PriceScheduleXp": {
        "type": "object",
        "example": {},
        "properties": {}
      },
      "ProductFacetXp": {
        "type": "object",
        "example": {"Options":[""],"ParentID":""},
        "properties": {
          "Options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ParentID": {
            "type": "string"
          }
        }
      },
      "ProductImage": {
        "type": "object",
        "example": {"URL":"","Tag":{"VariantID":"","Options":[""]}},
        "properties": {
          "URL": {
            "type": "string"
          },
          "Tag": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Tag"
              }
            ]
          }
        }
      },
      "ProductXp": {
        "type": "object",
        "example": {"Images":[{"URL":"","Tag":{"VariantID":"","Options":[""]}}],"Status":"Draft","HasVariants":false,"Note":"","Tax":{"Category":"","Code":"","Description":""},"UnitOfMeasure":{"Qty":0,"Unit":""},"ProductType":"Standard"},
        "properties": {
          "Images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImage"
            }
          },
          "Status": {
            "enum": [
              "Draft",
              "Published"
            ],
            "type": "string"
          },
          "HasVariants": {
            "type": "boolean"
          },
          "Note": {
            "type": "string",
            "maxLength": 500
          },
          "Tax": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TaxProperties"
              }
            ]
          },
          "UnitOfMeasure": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UnitOfMeasure"
              }
            ]
          },
          "ProductType": {
            "enum": [
              "Standard",
              "Quote"
            ],
            "type": "string"
          }
        }
      },
      "SpecOption": {
        "type": "object",
        "example": {"ID":"","Value":"","ListOrder":0,"IsOpenText":false,"PriceMarkupType":"NoMarkup","PriceMarkup":0},
        "properties": {
          "ID": {
            "type": "string"
          },
          "Value": {
            "type": "string"
          },
          "ListOrder": {
            "type": "integer",
            "format": "int32"
          },
          "IsOpenText": {
            "type": "boolean"
          },
          "PriceMarkupType": {
            "enum": [
              "NoMarkup",
              "AmountPerQuantity",
              "AmountTotal",
              "Percentage"
            ],
            "type": "string"
          },
          "PriceMarkup": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "SpecOptionXp": {
        "type": "object",
        "example": {"Description":"","SpecID":""},
        "properties": {
          "Description": {
            "type": "string"
          },
          "SpecID": {
            "type": "string"
          }
        }
      },
      "SpecUI": {
        "type": "object",
        "example": {"ControlType":"Text"},
        "properties": {
          "ControlType": {
            "enum": [
              "Text",
              "DropDown",
              "Checkbox",
              "Range"
            ],
            "type": "string"
          }
        }
      },
      "SpecXp": {
        "type": "object",
        "example": {"UI":{"ControlType":"Text"}},
        "properties": {
          "UI": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SpecUI"
              }
            ]
          }
        }
      },
...
      "Tag": {
        "type": "object",
        "example": {"VariantID":"","Options":[""]},
        "properties": {
          "VariantID": {
            "type": "string"
          },
          "Options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TaxProperties": {
        "type": "object",
        "example": {"Category":"","Code":"","Description":""},
        "properties": {
          "Category": {
            "type": "string"
          },
          "Code": {
            "type": "string"
          },
          "Description": {
            "type": "string",
            "maxLength": 100
          }
        }
      },
      "UnitOfMeasure": {
        "type": "object",
        "example": {"Qty":0,"Unit":""},
        "properties": {
          "Qty": {
            "type": "integer",
            "format": "int32"
          },
          "Unit": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/orchestration/{clientId}/catalog": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceCatalog"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostCatalog",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace catalog.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Catalog",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "CatalogAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/catalogproductassignment": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceCatalogAssignment"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostCatalogProductAssignment",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace catalog assignment.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Catalog Product Assignment",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "CatalogAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/priceschedule": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplacePriceSchedule"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostPriceSchedule",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace price schedule.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Price Schedule",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "PriceScheduleAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/product": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceProduct"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostProduct",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace product.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Product",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "ProductAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/productassignment": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceProductAssignment"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostProductAssignment",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace product assignment.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Product Assignment",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "ProductAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/productfacet": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceProductFacet"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostProductFacet",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace product facet.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Product Facet",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "ProductFacetAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/spec": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceSpec"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostSpec",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace spec.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Spec",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "ProductAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/specoption": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceSpecOption"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostSpecOption",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace spec option.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Spec Option",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "ProductAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/specproductassignment": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceSpecProductAssignment"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationProducts.PostSpecProductAssignment",
        "tags": [
          "Orchestration Products"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace spec product assignment.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Spec Product Assignment",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "ProductAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
...
  }
}